# openapi-updated.yaml
openapi: 3.0.3
info:
  title: Salla Journey Proxy API
  description: |
    A secure proxy service for Salla applications to communicate with Airflow APIs and handle webhook events.
    This API provides endpoints for proxying requests to Airflow, handling webhooks, and user management.
  version: 1.0.0
  contact:
    name: Salla Team
    url: https://salla.com
servers:
  - url: https://journey-proxy.salla.workers.dev
    description: Production server
  - url: http://localhost:8787
    description: Development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key with format "ApiKey {key}"
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Secret for webhook authentication
  
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "An error occurred"
        stack:
          type: string
          description: Only provided in development environment
    
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"
        message:
          type: string
          example: "Salla Journey Proxy API is running"
        version:
          type: string
          example: "1.0.0"
        timestamp:
          type: string
          format: date-time
          example: "2025-04-30T10:00:00Z"
    
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user123"
        password:
          type: string
          format: password
          example: "securepassword"
    
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            username:
              type: string
              example: "user123"
            email:
              type: string
              format: email
              example: "user@example.com"
    
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "user123"
        email:
          type: string
          format: email
          example: "user@example.com"
        createdAt:
          type: integer
          format: int64
          example: 1714482000
    
    ApiKeyRequest:
      type: object
      required:
        - keyName
      properties:
        keyName:
          type: string
          example: "Development API Key"
        expiresInDays:
          type: integer
          example: 30
    
    ApiKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        keyName:
          type: string
          example: "Development API Key"
        apiKey:
          type: string
          example: "11bf5b37-e0b8-42e0-8dcf-dc8c4aefc000"
        expiresAt:
          type: integer
          format: int64
          example: 1717160400
        createdAt:
          type: integer
          format: int64
          example: 1714482000

    ApiKeyListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        keyName:
          type: string
          example: "Development API Key"
        expiresAt:
          type: integer
          format: int64
          example: 1717160400
        createdAt:
          type: integer
          format: int64
          example: 1714482000
    
    WebhookResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Webhook received and processed"
        timestamp:
          type: string
          format: date-time
          example: "2025-04-30T10:05:00Z"

paths:
  /:
    get:
      summary: Health check endpoint
      description: Check if the API is running and healthy
      operationId: healthCheck
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  
  /users/login:
    post:
      summary: User login
      description: Authenticate user and get JWT token
      operationId: userLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Missing username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/me:
    get:
      summary: Get current user profile
      description: Get profile information for the authenticated user
      operationId: getUserProfile
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/api-keys:
    get:
      summary: List API keys
      description: Get all API keys for the current user
      operationId: listApiKeys
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: API keys retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyListItem'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create API key
      description: Create a new API key for the current user
      operationId: createApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyRequest'
      responses:
        '200':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '400':
          description: Missing key name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/api-keys/{keyId}:
    delete:
      summary: Delete API key
      description: Delete an API key by ID
      operationId: deleteApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the API key to delete
      responses:
        '200':
          description: API key deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /webhook:
    post:
      summary: Handle webhook
      description: Process incoming webhook events
      operationId: handleWebhook
      security:
        - WebhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # Webhook payloads can vary by event type
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /proxy:
    post:
      summary: Proxy POST request to Airflow API
      description: Forward POST request to Airflow API
      operationId: proxyPost
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: Proxy GET request to Airflow API
      description: Forward GET request to Airflow API
      operationId: proxyGet
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: params
          schema:
            type: object
            additionalProperties: true
          style: form
          explode: true
          description: Query parameters to forward to Airflow API
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Proxy PUT request to Airflow API
      description: Forward PUT request to Airflow API
      operationId: proxyPut
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      summary: Proxy PATCH request to Airflow API
      description: Forward PATCH request to Airflow API
      operationId: proxyPatch
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /proxy/{path}:
    post:
      summary: Proxy POST request with path to Airflow API
      description: Forward POST request with path to Airflow API
      operationId: proxyPostWithPath
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: Path to forward to Airflow API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: Proxy GET request with path to Airflow API
      description: Forward GET request with path to Airflow API
      operationId: proxyGetWithPath
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: Path to forward to Airflow API
        - in: query
          name: params
          schema:
            type: object
            additionalProperties: true
          style: form
          explode: true
          description: Query parameters to forward to Airflow API
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'